feat: enable more than one scheduler implementation
