+++ synapsys-examples/build.gradle.kts
@@ -10,6 +10,7 @@ repositories {
 
 kotlin {
     jvmToolchain(17)
+    //
 }
 
 dependencies {
